Goals given functionality. Created GoalModal and ProgressBar components. Shelf and Book components adjusted.
Goals can be added, edited, or deleted. ProgressBar shows progress of each goal. GoalModal is a popup for editing and adding a goal. Book and Shelf now take 'context' as a prop to show what page you are coming from. For now, only includes: if coming from the goal page, show 3/5 star rating. When a book is clicked it links to [book].tsx. All goal books currently use mock data.