CRUD operations
Crud operations added to clean up adding a book to a shelf using a main function with different match cases based on the shelf type.
Operation added to get the books in a shelf based on default shelf type.
Operation added to get the books in a custom shelf based on link table